pyproject.toml
[project]
name = "curso-python-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
